refactor : pr때 수정사항 추가
