+++ src/main/java/com/fantion/backend/common/dto/ReturnDto.java
@@ -1,12 +1,13 @@
-package com.fantion.backend.auction.dto;
+package com.fantion.backend.common.dto;
 
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import java.util.Map;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
+@Builder
 public class ReturnDto {
   private String message;
   private Map<String, Object> data;
